+++ packages/babel-helper-transform-fixture-test-runner/src/index.ts
@@ -637,6 +637,8 @@ const assertTest = function (
 ) {
   const expectStderr = opts.stderr.trim();
   stderr = stderr.trim();
+  stderr = stderr.replace(/\\\\/g, "/");
+  stderr = stderr.replace(/\\/g, "/");
 
   try {
     if (opts.stderr) {
@@ -656,6 +658,7 @@ const assertTest = function (
 
   const expectStdout = opts.stdout.trim();
   stdout = stdout.trim();
+  stdout = stdout.replace(/\\\\/g, "/");
   stdout = stdout.replace(/\\/g, "/");
 
   try {
@@ -728,6 +731,13 @@ export function buildParallelProcessTests(name: string, tests: ProcessTest[]) {
       }
     });
   };
+}
+
+const rootUrl = new URL("../../..", import.meta.url);
+function resolveRootDirOrRootUrlToken(arg: string) {
+  return arg
+    .replace("<rootDir>", fileURLToPath(rootUrl))
+    .replace("<rootUrl>", rootUrl.href.slice(0, -1));
 }
 
 export function buildProcessTests(
@@ -844,19 +854,17 @@ export function buildProcessTests(
           try {
             beforeHook(test, tmpLoc);
 
-            if (test.binLoc === undefined) {
-              throw new Error("test.binLoc is undefined");
+            let args = [];
+            if (opts.executor) {
+              args.push("--require", path.join(dirname, "./exit-loader.cjs"));
+            }
+            if (test.binLoc) {
+              args.push(test.binLoc);
             }
 
-            let args = opts.executor
-              ? [
-                  "--require",
-                  path.join(dirname, "./exit-loader.cjs"),
-                  test.binLoc,
-                ]
-              : [test.binLoc];
-
-            args = args.concat(opts.args);
+            args = args
+              .concat(opts.args)
+              .map(arg => resolveRootDirOrRootUrlToken(arg));
             const env = {
               ...process.env,
               FORCE_COLOR: "false",