+++ packages/babel-parser/test/fixtures/typescript/variable-declarator/invalid-definite-without-type/input.ts
@@ -1,0 +1,7 @@
+var w!;
+let x!;
+const y!;
+{
+  using z!;
+  await using za!;
+}